requirements.txt
geopy==2.4.1
python-dotenv==0.21.1
requests==2.31.0
pytest==8.3.2
click==8.1.7
filecache==0.81